feat: rewards sections unified, support for donors added
