+++ client/src/components/Projects/ProjectsList/ProjectsList.module.scss
@@ -37,5 +37,5 @@
   width: 100%;
   height: 0.1rem;
   background-color: $color-octant-grey1;
-  margin: 0 0 1.6rem;
+  margin: 1.6rem 0;
 }
