+++ client/src/components/Projects/ProjectsListItem/ProjectsListItem.module.scss
@@ -76,6 +76,7 @@
   }
 
   .projectRewards {
+    display: flex;
     flex: 1;
     padding: 0 $projectItemPadding;
   }
