+++ client/src/components/Projects/ProjectsListItem/ProjectsListItem.tsx
@@ -4,8 +4,7 @@ import { useNavigate } from 'react-router-dom';
 
 import ProjectsListSkeletonItem from 'components/Projects/ProjectsListSkeletonItem/ProjectsListSkeletonItem';
 import ButtonAddToAllocate from 'components/shared/ButtonAddToAllocate';
-import RewardsWithoutThreshold from 'components/shared/RewardsWithoutThreshold';
-import RewardsWithThreshold from 'components/shared/RewardsWithThreshold';
+import Rewards from 'components/shared/Rewards';
 import Description from 'components/ui/Description';
 import Img from 'components/ui/Img';
 import TinyLabel from 'components/ui/TinyLabel';
@@ -142,17 +141,8 @@ const ProjectsListItem: FC<ProjectsListItemProps> = ({
               text={introDescription!}
             />
           </div>
-          {!isEpoch1 && epoch && epoch < 4 && (
-            <RewardsWithThreshold
-              address={address}
-              className={styles.projectRewards}
-              epoch={epoch}
-              numberOfDonors={projectIpfsWithRewards.numberOfDonors}
-              totalValueOfAllocations={projectIpfsWithRewards.totalValueOfAllocations}
-            />
-          )}
-          {!isEpoch1 && (!epoch || epoch >= 4) && (
-            <RewardsWithoutThreshold
+          {!isEpoch1 && (
+            <Rewards
               address={address}
               className={styles.projectRewards}
               epoch={epoch}
