--- client/src/components/shared/RewardsWithThreshold/RewardsWithThreshold.module.scss
@@ -1,76 +1,0 @@
-.root {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  align-self: stretch;
-  justify-content: space-between;
-
-  .progressBar {
-    margin: 0.3rem 0;
-  }
-
-  .divider {
-    height: 0.2rem;
-    margin: 0.3rem 0;
-    width: 100%;
-    content: '';
-    background: $color-octant-grey3;
-  }
-
-  .sections {
-    margin-top: 1.2rem;
-    display: flex;
-    justify-content: space-between;
-    flex: 1;
-
-    .section {
-      font-weight: $font-weight-bold;
-      color: $color-octant-grey5;
-
-      .label,
-      .value {
-        &.isFetching {
-          @include skeleton();
-          text-indent: 1000%; // moves text outside of view.
-          white-space: nowrap;
-          overflow: hidden;
-          width: 8.8rem;
-        }
-      }
-
-      .label {
-        font-size: $font-size-12;
-        height: 1.6rem;
-        margin-bottom: 0.4rem;
-        display: flex;
-        align-items: center;
-      }
-
-      &.leftSection {
-        text-align: left;
-
-        .label {
-          justify-content: flex-start;
-        }
-
-        .value {
-          &.greenValue {
-            color: $color-octant-green;
-          }
-
-          &.redValue {
-            color: $color-octant-orange;
-          }
-        }
-      }
-
-      &.rightSection {
-        text-align: right;
-
-        .label {
-          justify-content: flex-end;
-        }
-      }
-    }
-  }
-}
