--- client/src/components/shared/RewardsWithoutThreshold/index.tsx
@@ -1,2 +1,0 @@
-// eslint-disable-next-line no-restricted-exports
-export { default } from './RewardsWithoutThreshold';
