--- client/src/components/shared/RewardsWithoutThreshold/utils.ts
@@ -1,8 +1,0 @@
-export function getProgressPercentage(
-  totalValueOfAllocations: bigint,
-  cutOffValue: bigint,
-): number {
-  return totalValueOfAllocations > 0 && cutOffValue > 0
-    ? Number((totalValueOfAllocations * 100n) / cutOffValue)
-    : 0;
-}
